requirements.txt
appdirs==1.4.4
attrs==21.2.0
backports.zoneinfo==0.2.1
black==21.7b0
cffi==1.14.6
click==8.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
matlabengineforpython===R2021a
MIDIUtil==1.2.1
mido==1.2.10
mypy-extensions==0.4.3
numpy==1.21.1
pandas==1.3.1
pathspec==0.9.0
pretty-midi==0.2.9
pycparser==2.20
PyGuitarPro==0.8
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
rpy2==3.4.5
scipy==1.7.1
six==1.16.0
tomli==1.2.1
tzlocal==3.0
